Add events for service start and stop
Add two new events that are sent to the event hook when a service is
started and when it is stopped. These two hooks provide observability
during normal operations. Note that the stop event may be followed by a
terminate or panic event if the service was not stopped permanently.

fixes #76